Remove routing
Handling page state in the app-db seems adequate. Using secretary for client side routing and having that pretty urls was nice but at this point the benefits that brings (users can navigate to different sections of the app directly) doesn't out way the cost of having to manage user navigation with syncing what is actually in the URL bar with history modification (ugly side-effects and messiness). Pretty URLs can be added later.