Expand Vagrantfile rsync exclusions to fix some incorrect file syncing
Noteably this should fix "Exception one of :jdbc-url, :adapter, :datasource, or :datasource-classname is required to initialize the connection! conman.core/make-config (core.clj:56)" bug which seems to be caused by a the .lein-env getting written at REPL start then a developer remember to actually enable rsync-auto and the file getting overwritten or deleted.